submitTurn shouldn't be a "move"
As of now, the framework doesn't allow me to validate a move on the server side when a client calls endTurn,
which is why I made submitTurn a real "move" to begin with.
The client should make requests, not decisions. Eventually we'll need the framework to account for this.
But for now, making submitTurn a "move" is just making things hard to maintain.